Revert "Merge tag 'v3.1' into cm-10.1"
This reverts commit a25ee1785227ae35c671e3b728b7ac0196dbf62a, reversing
changes made to 064e52774a36c757aa06a953ccde4fcc254b3817.

+++ fs/autofs4/waitq.c
@@ -104,7 +104,7 @@ static void autofs4_notify_daemon(struct autofs_sb_info *sbi,
 	size_t pktsz;
 
 	DPRINTK("wait id = 0x%08lx, name = %.*s, type=%d",
-		(unsigned long) wq->wait_queue_token, wq->name.len, wq->name.name, type);
+		wq->wait_queue_token, wq->name.len, wq->name.name, type);
 
 	memset(&pkt,0,sizeof pkt); /* For security reasons */
